pyproject.toml
[project]
name = "zmq-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "msgpack>=1.1.2",
    "pyzmq>=27.1.0",
]
